Avoid lookup in d_revalidate()
This patch optimize lookups based upon VFS patch which allows
positive dentry to be passed on to atomic_open().

In fuse_dentry_revalidate(), we set dentry flag
DCACHE_ATOMIC_OPEN if open atomic is implemented by
underlying File System. This flag let the positive dentry
come into file system specific atomic open through VFS
atomic_open(). FS specific atomic open might create or
open the file.

If file was newly created, we invalidate the positive dentry.
New dentry and a new inode is allocated and both linked.
File is now opened for usage.

If file was not created, we revalidate the inode. If inode is
stale, current inode is marked as bad. And new inode is allocated
and linked to new dentry(old dentry invalidated). In case of
inode attributes differing with fresh attr, we allocate new
dentry and hook current inode to it and open the file.

For negative dentry, FS just allocate new inode and hook it onto
passed entry from VFS and open the file.