Implement open atomic
When we go for creating a file, we trigger
a lookup into FUSE USER space. It is very  much likely
that file does not exist yet as O_CREAT is passed to
open(). This extra lookup call can be avoided.

Also we trigger a lookup when we open a file (dentry
timedout in d_revalidate(), this lookup can also be
avoided with this new open atomic implementation.

Here is how current FUSE create works:

A. Looks up dentry if d_in_lookup() is set.
B. If dentry is positive or O_CREAT is not set, return.
C. If server supports atomic create + open, use it to create
   and open the file.
D. If server does not support atomic create + open, just create file
   using "mknod" and return. VFS will take care of opening the file.

And here it is how current FUSE open works:

A. Looks up dentry if d_in_lookup() is set.
B. If dentry is positive or O_CREAT is not set, go to VFS for open.
C. If server supports atomic create + open, use it to create
   and open the file.

Here is how the proposed patch would work for CREATE and OPEN:

A. Skip lookup if open atomic is supported and file is being
   created/opened.
B. If open atomic is not implemented then fall back to old way
   of creating and opening the file.

It is expected that USER SPACE creates and open the file as
applicable (depending upon passed open flags) and fills in the
attributes which are then used to make inode stand/revalidate in
the kernel cache. If file was newly created in USER SPACE
then it must be indicated in 'struct fuse_file_info' by setting
a bit 'file_created' which is used by libfuse to send flags back
to the Fuse kernel to indicate that that file was indeed newly
created. These flags are used by kernel to indicate changes in
parent directory and set CREATE flags on the file.

Fuse kernel automatically detects if open atomic is implemented
by libfuse/USER SPACE. And depending upon the outcome of this
check all further create and open calls are routed accordingly.

If libfuse/USER SPACE has not implemented the open atomic then
there is no change in default behaviour i.e lookups are not
optimized.